Revert "feat: #469 - ClassDeclaration - Add "Extract interface""
This reverts commit 440f8389c1265c1dab9d24a55b57d446693f0614.
