+++ src/compiler/ast/base/BodiedNode.ts
@@ -4,7 +4,7 @@ import { Constructor, WriterFunction } from "../../../types";
 import { ts } from "../../../typescript";
 import { callBaseSet } from "../callBaseSet";
 import { Node } from "../common";
-import { getBodyTextWithoutLeadingIndentation, setBodyTextForNode } from "./helpers";
+import { getBodyTextForStructure, setBodyTextForNode } from "./helpers";
 import { callBaseGetStructure } from "../callBaseGetStructure";
 
 export type BodiedNodeExtensionType = Node<ts.Node & { body: ts.Node; }>;
@@ -19,10 +19,6 @@ export interface BodiedNode {
      * @param textOrWriterFunction - Text or writer function to set as the body.
      */
     setBodyText(textOrWriterFunction: string | WriterFunction): this;
-    /**
-     * Gets the body text without leading whitespace, leading indentation, or trailing whitespace.
-     */
-    getBodyText(): string;
 }
 
 export function BodiedNode<T extends Constructor<BodiedNodeExtensionType>>(Base: T): Constructor<BodiedNode> & T {
@@ -50,13 +46,9 @@ export function BodiedNode<T extends Constructor<BodiedNodeExtensionType>>(Base:
             return this;
         }
 
-        getBodyText() {
-            return getBodyTextWithoutLeadingIndentation(this.getBody());
-        }
-
         getStructure() {
             return callBaseGetStructure<BodiedNodeStructure>(Base.prototype, this, {
-                bodyText: this.getBodyText()
+                bodyText: getBodyTextForStructure(this.getBody())
             });
         }
     };
