+++ src/compiler/ast/base/helpers/getBodyTextForStructure.ts
@@ -4,7 +4,7 @@ import { Node } from "../../common";
 /**
  * @internal
  */
-export function getBodyTextWithoutLeadingIndentation(body: Node) {
+export function getBodyTextForStructure(body: Node) {
     const sourceFile = body.sourceFile;
     const textArea = body.getChildSyntaxList() || body; // arrow functions don't have a syntax list
     const startPos = textArea.getNonWhitespaceStart();
