+++ src/compiler/ast/class/base/index.ts
@@ -1,2 +1,1 @@
 export * from "./AbstractableNode";
-export * from "./ClassLikeDeclarationBase";
