+++ src/compiler/ast/class/index.ts
@@ -1,6 +1,5 @@
 export * from "./base";
 export * from "./ClassDeclaration";
-export * from "./ClassExpression";
 export * from "./ConstructorDeclaration";
 export * from "./GetAccessorDeclaration";
 export * from "./MethodDeclaration";
