+++ src/compiler/ast/kindToNodeMappings.ts
@@ -21,7 +21,6 @@ export interface KindToNodeMappings {
     [SyntaxKind.CaseClause]: compiler.CaseClause;
     [SyntaxKind.CatchClause]: compiler.CatchClause;
     [SyntaxKind.ClassDeclaration]: compiler.ClassDeclaration;
-    [SyntaxKind.ClassExpression]: compiler.ClassExpression;
     [SyntaxKind.Constructor]: compiler.ConstructorDeclaration;
     [SyntaxKind.ConstructorType]: compiler.ConstructorTypeNode;
     [SyntaxKind.ConstructSignature]: compiler.ConstructSignatureDeclaration;
@@ -176,7 +175,6 @@ export interface KindToExpressionMappings {
     [SyntaxKind.AwaitExpression]: compiler.AwaitExpression;
     [SyntaxKind.BinaryExpression]: compiler.BinaryExpression;
     [SyntaxKind.CallExpression]: compiler.CallExpression;
-    [SyntaxKind.ClassExpression]: compiler.ClassExpression;
     [SyntaxKind.CommaListExpression]: compiler.CommaListExpression;
     [SyntaxKind.ConditionalExpression]: compiler.ConditionalExpression;
     [SyntaxKind.DeleteExpression]: compiler.DeleteExpression;
