+++ src/tests/compiler/base/questionTokenableNodeTests.ts
@@ -1,61 +1,59 @@
 ﻿import { expect } from "chai";
-import { ClassDeclaration, PropertyDeclaration, QuestionTokenableNode } from "../../../compiler";
+import { ParameterDeclaration, ClassDeclaration, PropertyDeclaration, QuestionTokenableNode } from "../../../compiler";
+import * as errors from "../../../errors";
+import { SyntaxKind } from "../../../typescript";
 import { QuestionTokenableNodeStructure } from "../../../structures";
-import { getInfoFromText } from "../testHelpers";
+import { getInfoFromText, getInfoFromTextWithDescendant } from "../testHelpers";
 
 describe(nameof(QuestionTokenableNode), () => {
-    function getInfoWithFirstMember(text: string) {
+    function getInfoWithFirstPropertyFromText(text: string) {
         const result = getInfoFromText<ClassDeclaration>(text);
-        return {...result, firstMember: result.firstChild.getMembers()[0] as QuestionTokenableNode };
+        return {...result, firstProperty: result.firstChild.getInstanceProperties()[0] as PropertyDeclaration };
     }
 
     describe(nameof<QuestionTokenableNode>(d => d.hasQuestionToken), () => {
         function doTest(text: string, value: boolean) {
-            const {firstMember} = getInfoWithFirstMember(text);
-            expect(firstMember.hasQuestionToken()).to.equal(value);
+            const {firstProperty} = getInfoWithFirstPropertyFromText(text);
+            expect(firstProperty.hasQuestionToken()).to.equal(value);
         }
 
-        it("should have when property with one", () => {
+        it("should have a question token when has one", () => {
             doTest("class MyClass { prop?: string; }", true);
         });
 
-        it("should have when a method with one", () => {
-            doTest("declare class MyClass { myMethod?(): string; }", true);
-        });
-
-        it("should not have when not has one", () => {
+        it("should not have a question token when not has one", () => {
             doTest("class MyClass { prop: string; }", false);
         });
     });
 
     describe(nameof<QuestionTokenableNode>(d => d.getQuestionTokenNode), () => {
         it("should be get the question token node", () => {
-            const {firstMember} = getInfoWithFirstMember("class MyClass { prop?: string; }");
-            expect(firstMember.getQuestionTokenNode()!.getText()).to.equal("?");
+            const {firstProperty} = getInfoWithFirstPropertyFromText("class MyClass { prop?: string; }");
+            expect(firstProperty.getQuestionTokenNode()!.getText()).to.equal("?");
         });
 
         it("should be undefined when not optional", () => {
-            const {firstMember} = getInfoWithFirstMember("class MyClass { prop: string;} ");
-            expect(firstMember.getQuestionTokenNode()).to.be.undefined;
+            const {firstProperty} = getInfoWithFirstPropertyFromText("class MyClass { prop: string;} ");
+            expect(firstProperty.getQuestionTokenNode()).to.be.undefined;
         });
     });
 
     describe(nameof<QuestionTokenableNode>(d => d.getQuestionTokenNodeOrThrow), () => {
         it("should be get the question token node", () => {
-            const {firstMember} = getInfoWithFirstMember("class MyClass {\nprop?: string;}\n");
-            expect(firstMember.getQuestionTokenNodeOrThrow().getText()).to.equal("?");
+            const {firstProperty} = getInfoWithFirstPropertyFromText("class MyClass {\nprop?: string;}\n");
+            expect(firstProperty.getQuestionTokenNodeOrThrow().getText()).to.equal("?");
         });
 
         it("should throw when not optional", () => {
-            const {firstMember} = getInfoWithFirstMember("class MyClass {\nprop: string;}\n");
-            expect(() => firstMember.getQuestionTokenNodeOrThrow()).to.throw();
+            const {firstProperty} = getInfoWithFirstPropertyFromText("class MyClass {\nprop: string;}\n");
+            expect(() => firstProperty.getQuestionTokenNodeOrThrow()).to.throw();
         });
     });
 
     describe(nameof<QuestionTokenableNode>(d => d.setHasQuestionToken), () => {
         function doTest(startText: string, value: boolean, expected: string) {
-            const {firstMember, sourceFile} = getInfoWithFirstMember(startText);
-            firstMember.setHasQuestionToken(value);
+            const {firstProperty, sourceFile} = getInfoWithFirstPropertyFromText(startText);
+            firstProperty.setHasQuestionToken(value);
             expect(sourceFile.getFullText()).to.be.equal(expected);
         }
 
@@ -82,24 +80,12 @@ describe(nameof(QuestionTokenableNode), () => {
         it("should be set as optional when has no type nor semi-colon", () => {
             doTest("class MyClass { prop }", true, "class MyClass { prop? }");
         });
-
-        it("should set when a method", () => {
-            doTest("declare class MyClass { method(): string; }", true, "declare class MyClass { method?(): string; }");
-        });
-
-        it("should set when a method with type parameters", () => {
-            doTest("declare class MyClass { method<T>(): string; }", true, "declare class MyClass { method?<T>(): string; }");
-        });
-
-        it("should remove when a method", () => {
-            doTest("declare class MyClass { method?(): string; }", false, "declare class MyClass { method(): string; }");
-        });
     });
 
     describe(nameof<PropertyDeclaration>(p => p.set), () => {
         function doTest(startCode: string, structure: QuestionTokenableNodeStructure, expectedCode: string) {
-            const {firstMember, sourceFile} = getInfoWithFirstMember(startCode);
-            (firstMember as PropertyDeclaration).set(structure);
+            const {firstProperty, sourceFile} = getInfoWithFirstPropertyFromText(startCode);
+            firstProperty.set(structure);
             expect(sourceFile.getText()).to.equal(expectedCode);
         }
 
@@ -118,8 +104,8 @@ describe(nameof(QuestionTokenableNode), () => {
 
     describe(nameof<PropertyDeclaration>(p => p.getStructure), () => {
         function doTest(startCode: string, hasToken: boolean) {
-            const {firstMember} = getInfoWithFirstMember(startCode);
-            expect((firstMember as PropertyDeclaration).getStructure().hasQuestionToken).to.equal(hasToken);
+            const {firstProperty, sourceFile} = getInfoWithFirstPropertyFromText(startCode);
+            expect(firstProperty.getStructure().hasQuestionToken).to.equal(hasToken);
         }
 
         it("should be false when not has one", () => {
