+++ src/utils/TypeGuards.ts
@@ -30,7 +30,7 @@ export class TypeGuards {
      * Gets if the node has a name.
      * @param node - Node to check.
      */
-    static hasName(node: compiler.Node): node is compiler.Node & { getName(): string; getNameNode(): compiler.Node; } {
+    static hasName(node: compiler.Node): node is compiler.Node & { getName(): string; } {
         // this method is manually maintained
         if ((node as any).getName == null)
             return false;
@@ -55,7 +55,6 @@ export class TypeGuards {
     static isAbstractableNode(node: compiler.Node): node is compiler.AbstractableNode & compiler.AbstractableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
             case SyntaxKind.PropertyDeclaration:
@@ -372,28 +371,6 @@ export class TypeGuards {
     }
 
     /**
-     * Gets if the node is a ClassExpression.
-     * @param node - Node to check.
-     */
-    static isClassExpression(node: compiler.Node): node is compiler.ClassExpression {
-        return node.getKind() === SyntaxKind.ClassExpression;
-    }
-
-    /**
-     * Gets if the node is a ClassLikeDeclarationBase.
-     * @param node - Node to check.
-     */
-    static isClassLikeDeclarationBase(node: compiler.Node): node is compiler.ClassLikeDeclarationBase & compiler.ClassLikeDeclarationBaseExtensionType {
-        switch (node.getKind()) {
-            case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    /**
      * Gets if the node is a CommaListExpression.
      * @param node - Node to check.
      */
@@ -472,7 +449,6 @@ export class TypeGuards {
     static isDecoratableNode(node: compiler.Node): node is compiler.DecoratableNode & compiler.DecoratableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
             case SyntaxKind.PropertyDeclaration:
@@ -619,7 +595,6 @@ export class TypeGuards {
             case SyntaxKind.StringKeyword:
             case SyntaxKind.SymbolKeyword:
             case SyntaxKind.UndefinedKeyword:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.AsExpression:
             case SyntaxKind.AwaitExpression:
@@ -850,7 +825,6 @@ export class TypeGuards {
     static isHeritageClauseableNode(node: compiler.Node): node is compiler.HeritageClauseableNode & compiler.HeritageClauseableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.InterfaceDeclaration:
                 return true;
             default:
@@ -879,13 +853,7 @@ export class TypeGuards {
      * @param node - Node to check.
      */
     static isImplementsClauseableNode(node: compiler.Node): node is compiler.ImplementsClauseableNode & compiler.ImplementsClauseableNodeExtensionType {
-        switch (node.getKind()) {
-            case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
-                return true;
-            default:
-                return false;
-        }
+        return node.getKind() === SyntaxKind.ClassDeclaration;
     }
 
     /**
@@ -1146,7 +1114,6 @@ export class TypeGuards {
     static isJSDocableNode(node: compiler.Node): node is compiler.JSDocableNode & compiler.JSDocableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Constructor:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
@@ -1306,7 +1273,6 @@ export class TypeGuards {
      */
     static isLeftHandSideExpression(node: compiler.Node): node is compiler.LeftHandSideExpression {
         switch (node.getKind()) {
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.CallExpression:
             case SyntaxKind.ElementAccessExpression:
@@ -1404,7 +1370,6 @@ export class TypeGuards {
      */
     static isMemberExpression(node: compiler.Node): node is compiler.MemberExpression {
         switch (node.getKind()) {
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.ElementAccessExpression:
             case SyntaxKind.ImportKeyword:
@@ -1465,7 +1430,6 @@ export class TypeGuards {
     static isModifierableNode(node: compiler.Node): node is compiler.ModifierableNode & compiler.ModifierableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Constructor:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
@@ -1511,7 +1475,6 @@ export class TypeGuards {
     static isNameableNode(node: compiler.Node): node is compiler.NameableNode & compiler.NameableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.FunctionDeclaration:
             case SyntaxKind.FunctionExpression:
                 return true;
@@ -1765,7 +1728,6 @@ export class TypeGuards {
      */
     static isPrimaryExpression(node: compiler.Node): node is compiler.PrimaryExpression {
         switch (node.getKind()) {
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.ImportKeyword:
             case SyntaxKind.MetaProperty:
@@ -1858,7 +1820,6 @@ export class TypeGuards {
      */
     static isQuestionTokenableNode(node: compiler.Node): node is compiler.QuestionTokenableNode & compiler.QuestionTokenableNodeExtensionType {
         switch (node.getKind()) {
-            case SyntaxKind.MethodDeclaration:
             case SyntaxKind.PropertyDeclaration:
             case SyntaxKind.Parameter:
             case SyntaxKind.MethodSignature:
@@ -1895,7 +1856,6 @@ export class TypeGuards {
         switch (node.getKind()) {
             case SyntaxKind.BindingElement:
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
             case SyntaxKind.PropertyDeclaration:
@@ -1941,7 +1901,6 @@ export class TypeGuards {
         switch (node.getKind()) {
             case SyntaxKind.BindingElement:
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
             case SyntaxKind.PropertyDeclaration:
@@ -2287,7 +2246,6 @@ export class TypeGuards {
     static isTextInsertableNode(node: compiler.Node): node is compiler.TextInsertableNode & compiler.TextInsertableNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Constructor:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
@@ -2467,7 +2425,6 @@ export class TypeGuards {
     static isTypeParameteredNode(node: compiler.Node): node is compiler.TypeParameteredNode & compiler.TypeParameteredNodeExtensionType {
         switch (node.getKind()) {
             case SyntaxKind.ClassDeclaration:
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Constructor:
             case SyntaxKind.GetAccessor:
             case SyntaxKind.MethodDeclaration:
@@ -2520,7 +2477,6 @@ export class TypeGuards {
      */
     static isUnaryExpression(node: compiler.Node): node is compiler.UnaryExpression {
         switch (node.getKind()) {
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.AwaitExpression:
             case SyntaxKind.CallExpression:
@@ -2612,7 +2568,6 @@ export class TypeGuards {
      */
     static isUpdateExpression(node: compiler.Node): node is compiler.UpdateExpression {
         switch (node.getKind()) {
-            case SyntaxKind.ClassExpression:
             case SyntaxKind.Identifier:
             case SyntaxKind.CallExpression:
             case SyntaxKind.ElementAccessExpression:
